requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
mongomock==4.2.0
gunicorn==22.0.0
python-dotenv==1.0.0

